pyproject.toml
[project]
name = "eafawproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "crewai>=1.2.1",
    "crewai-tools>=1.2.1",
    "fastapi>=0.120.1",
    "groq>=0.33.0",
    "langchain>=1.0.2",
    "logger-py>=1.0.0",
    "numpy>=2.3.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
]
